requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==0.28.1
requests==2.31.0
bleach==6.1.0